mv listbot discord
