+++ src/index.js
@@ -10,7 +10,7 @@ const csrfM = require('./csrf');
 const config = require('config');
 const botM = require('./bot');
 const listM = require('./list');
-const bot = require('./listbot');
+const bot = require('./discord');
 const themeM = require('./theme');
 const express = require('express');
 const discM = require('./discord');
