+++ src/user.js
@@ -1,5 +1,5 @@
 const config = require('config');
-const bot = require('./listbot');
+const bot = require('./discord');
 
 const userSetup = (req, res, next) => {
 	if (req.user && bot.client.startTime) {
